Update README.md with obsolescence message
I stopped using habitrpg around the time the API had a breaking change, and never circled back to update this project. Since it was only here for me to tinker with the data, I figure it's best to archive the project so anyone googling for visualizations don't accidentally try to use it.